[#19] refactor: 기록 생성 api record.ts 파일로 이동
